Revert "Migrate modules 4 (#1518)"
This reverts commit 8968bcb302a2fd92332d79442bb3e189b2dc58de.
