+++ public/modules/ui/editors.js
@@ -21,8 +21,8 @@ function clicked() {
   if (!ancestor) return;
 
   if (grand.id === "emblems") editEmblem();
-  else if (parent.id === "rivers") window.Controllers.RiverEditor.open(el.id);
-  else if (grand.id === "routes") window.Controllers.RouteEditor.open(el.id);
+  else if (parent.id === "rivers") editRiver(el.id);
+  else if (grand.id === "routes") editRoute(el.id);
   else if (ancestor.id === "labels" && el.tagName === "tspan") editLabel();
   else if (grand.id === "burgLabels" || grand.id === "burgIcons") window.Controllers.BurgEditor.open(el.dataset.id);
   else if (parent.id === "ice") editIce(el);
